chore: refactoring tests
