+++ api/app/models/utils.py
@@ -1,4 +1,4 @@
-from typing import Any
+from typing import Any, Optional
 
 from app.models.inferred_schema_dto import SupportedTypes
 
@@ -11,7 +11,7 @@ def is_number_or_string(value: SupportedTypes):
     return value in (SupportedTypes.int, SupportedTypes.float, SupportedTypes.string)
 
 
-def is_optional_float(value: Any) -> bool:
+def is_optional_float(value: Optional[SupportedTypes] = None) -> bool:
     return value in (None, SupportedTypes.float)
