+++ api/tests/validation/model_type_validator_test.py
@@ -1,75 +1,8 @@
 from pydantic import ValidationError
 import pytest
 
-from app.models.inferred_schema_dto import SupportedTypes
-from app.models.model_dto import (
-    ColumnDefinition,
-    DataType,
-    Granularity,
-    ModelIn,
-    ModelType,
-    OutputType,
-)
-
-
-def get_model_sample_wrong(fail_field: str, model_type: ModelType):
-    prediction = None
-    prediction_proba = None
-    if fail_field == 'outputs.prediction' and model_type == ModelType.BINARY:
-        prediction = ColumnDefinition(name='pred1', type=SupportedTypes.string)
-    elif fail_field == 'outputs.prediction' and model_type == ModelType.MULTI_CLASS:
-        prediction = ColumnDefinition(name='pred1', type=SupportedTypes.datetime)
-    elif fail_field == 'outputs.prediction' and model_type == ModelType.REGRESSION:
-        prediction = ColumnDefinition(name='pred1', type=SupportedTypes.string)
-    else:
-        prediction = ColumnDefinition(name='pred1', type=SupportedTypes.int)
-
-    if (
-        fail_field == 'outputs.prediction_proba'
-        and model_type == ModelType.BINARY
-        or fail_field == 'outputs.prediction_proba'
-        and model_type == ModelType.MULTI_CLASS
-    ):
-        prediction_proba = ColumnDefinition(name='prob1', type=SupportedTypes.int)
-    elif (
-        fail_field == 'outputs.prediction_proba' and model_type == ModelType.REGRESSION
-    ):
-        prediction_proba = ColumnDefinition(name='prob1', type=SupportedTypes.float)
-    else:
-        prediction_proba = ColumnDefinition(name='prob1', type=SupportedTypes.float)
-
-    target: ColumnDefinition = None
-    if fail_field == 'target' and model_type == ModelType.BINARY:
-        target = ColumnDefinition(name='target1', type=SupportedTypes.string)
-    elif fail_field == 'target' and model_type == ModelType.MULTI_CLASS:
-        target = ColumnDefinition(name='target1', type=SupportedTypes.datetime)
-    elif fail_field == 'target' and model_type == ModelType.REGRESSION:
-        target = ColumnDefinition(name='target1', type=SupportedTypes.string)
-    else:
-        target = ColumnDefinition(name='target1', type=SupportedTypes.int)
-
-    timestamp: ColumnDefinition = None
-    if fail_field == 'timestamp':
-        timestamp = ColumnDefinition(name='timestamp', type=SupportedTypes.string)
-    else:
-        timestamp = ColumnDefinition(name='timestamp', type=SupportedTypes.datetime)
-
-    return {
-        'name': 'model_name',
-        'model_type': model_type,
-        'data_type': DataType.TEXT,
-        'granularity': Granularity.DAY,
-        'features': [ColumnDefinition(name='feature1', type=SupportedTypes.string)],
-        'outputs': OutputType(
-            prediction=prediction,
-            prediction_proba=prediction_proba,
-            output=[ColumnDefinition(name='output1', type=SupportedTypes.string)],
-        ),
-        'target': target,
-        'timestamp': timestamp,
-        'frameworks': None,
-        'algorithm': None,
-    }
+from app.models.model_dto import ModelIn, ModelType
+from tests.commons.modelin_factory import get_model_sample_wrong
 
 
 def test_timestamp_not_datetime():
@@ -101,7 +34,7 @@ def test_target_for_multiclass():
 
 
 def test_target_for_regression():
-    """Tests that for ModelType.REGRESSION: target must be a number or string."""
+    """Tests that for ModelType.REGRESSION: target must be a number."""
     with pytest.raises(ValidationError) as excinfo:
         model_data = get_model_sample_wrong('target', ModelType.REGRESSION)
         ModelIn.model_validate(ModelIn(**model_data))
@@ -117,7 +50,7 @@ def test_prediction_for_binary():
 
 
 def test_prediction_for_multiclass():
-    """Tests that for ModelType.MULTI_CLASS: prediction must be a number."""
+    """Tests that for ModelType.MULTI_CLASS: prediction must be a number or string."""
     with pytest.raises(ValidationError) as excinfo:
         model_data = get_model_sample_wrong('outputs.prediction', ModelType.MULTI_CLASS)
         ModelIn.model_validate(ModelIn(**model_data))
@@ -162,7 +95,7 @@ def test_prediction_proba_for_multiclass():
 
 
 def test_prediction_proba_for_regression():
-    """Tests that for ModelType.REGRESSION: prediction_proba must be a number."""
+    """Tests that for ModelType.REGRESSION: prediction_proba must be None."""
     with pytest.raises(ValidationError) as excinfo:
         model_data = get_model_sample_wrong(
             'outputs.prediction_proba', ModelType.REGRESSION
